backend/requirements.txt
fastapi
uvicorn
python-dotenv
httpx
pydantic
sqlalchemy
psycopg2-binary
openai-whisper
google-generativeai
python-multipart
